+++ openstack/autoscaling/v1/configurations/create.go
@@ -1,0 +1,15 @@
+package configurations
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Create(client *golangsdk.ServiceClient, opts CreateOptsBuilder) (r CreateResult) {
+	b, err := opts.ToConfigurationCreateMap()
+	if err != nil {
+		r.Err = err
+		return
+	}
+	_, r.Err = client.Post(client.ServiceURL("scaling_configuration"), b, &r.Body, &golangsdk.RequestOpts{
+		OkCodes: []int{200},
+	})
+	return
+}
