+++ openstack/autoscaling/v1/configurations/delete.go
@@ -1,0 +1,8 @@
+package configurations
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Delete(client *golangsdk.ServiceClient, id string) (r DeleteResult) {
+	_, r.Err = client.Delete(client.ServiceURL("scaling_configuration", id), nil)
+	return
+}
