+++ openstack/autoscaling/v1/configurations/list.go
@@ -1,0 +1,21 @@
+package configurations
+
+import (
+	"github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
+)
+
+func List(client *golangsdk.ServiceClient, opts ListOptsBuilder) pagination.Pager {
+	url := client.ServiceURL("scaling_configuration")
+	if opts != nil {
+		query, err := opts.ToConfigurationListQuery()
+		if err != nil {
+			return pagination.Pager{Err: err}
+		}
+		url += query
+	}
+
+	return pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
+		return ConfigurationPage{pagination.SinglePageBase(r)}
+	})
+}
