--- openstack/autoscaling/v1/configurations/urls.go
@@ -1,23 +1,0 @@
-package configurations
-
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-)
-
-const resourcePath = "scaling_configuration"
-
-func createURL(client *golangsdk.ServiceClient) string {
-	return client.ServiceURL(resourcePath)
-}
-
-func getURL(client *golangsdk.ServiceClient, id string) string {
-	return client.ServiceURL(resourcePath, id)
-}
-
-func deleteURL(client *golangsdk.ServiceClient, id string) string {
-	return client.ServiceURL(resourcePath, id)
-}
-
-func listURL(client *golangsdk.ServiceClient) string {
-	return client.ServiceURL(resourcePath)
-}
